Element Classes now support creating New Elements
Updated the Element sub-Classes to support element creation (by providing relevant attributes).

Minor modification to Events.js to show the selection area after element creation (the functionality will be moved from the Draw class eventually).

Commented out code that is no longer used from Draw.js

Commented out code that is no longer used from SVG.js
