+++ App/CanvasElements/Elements/Element/Circle.js
@@ -8,15 +8,18 @@ class Circle extends Element {
 	#r;
 
 	constructor() {
-		super();
-     	var radius = Math.sqrt(Math.pow(cache.stop[0] - cache.start[0], 2) + Math.pow(cache.stop[1] - cache.start[1], 2));
-		svg.new('circle', {
+		super(); // super must be always be called in a Sub-Class in JavaScript
+		// This constructor might be used in the future, although it's not guaranteed
+	}
+
+	createAttr() {
+		var radius = Math.sqrt(Math.pow(cache.stop[0] - cache.start[0], 2) + Math.pow(cache.stop[1] - cache.start[1], 2));
+		
+		return {
 			'cx': cache.start[0],
 			'cy': cache.start[1],
-			'r': radius,
-			'fill': tool.fill
-		});
-		
+			'r': radius
+		}
 	}
 
 	parseAttr(ele) {
