+++ App/CanvasElements/Main/SVG.js
@@ -12,95 +12,27 @@ var translateY = 0;
 var svg = {
     numID: 0, // each time a new element is added, the ID is incremented
     created: false,
-    new: function(ele, attr, id) {
-      if (svg.created == false) {
-        if (!id) {
-          id = this.numID;
-          this.numID++;
-        }
-        // the below code simply append to the HTML Document
-        $('#editor').html($('#editor').html() + '<' + ele + ' id=' + id + '/>');
-        this.id = '#' + id;
-        cache.ele = id;
-        this.created = true;
-        this.finished = false;
-      }
-      attr['stroke-width'] = tool.strokeWidth;
-      attr.stroke = tool.stroke;
-      if (ele != 'line' && ele != 'polyline') {
-        attr['paint-order'] = tool.paintOrder;
-      }
-      $(this.id).attr(attr);
-    },
-    storeAttr: function() {
-      cache.origSelectArea = {
-        x: $('.selection')[0].getBoundingClientRect().left,
-        y: $('.selection')[0].getBoundingClientRect().top,
-        width: $('.selection')[0].getBoundingClientRect().width,
-        height: $('.selection')[0].getBoundingClientRect().height
-      }
-      this.type = cache.ele[0].outerHTML.split(' ')[0].replace('<','');
-      switch (this.type) { // Stores element-specific data for later manipulation
-        case 'line':
-          this.line = {
-            x1: parseFloat(cache.ele.attr('x1')),
-            x2: parseFloat(cache.ele.attr('x2')),
-            y1: parseFloat(cache.ele.attr('y1')),
-            y2: parseFloat(cache.ele.attr('y2'))
-          }
-          break;
-        case 'rect':
-          this.rect = {
-            x: parseFloat(cache.ele.attr('x')),
-            y: parseFloat(cache.ele.attr('y')),
-            width: parseFloat(cache.ele.attr('width')),
-            height: parseFloat(cache.ele.attr('height'))
-          }
-          break;
-        case 'ellipse':
-          this.ellipse = {
-            cx: parseFloat(cache.ele.attr('cx')),
-            cy: parseFloat(cache.ele.attr('cy')),
-            rx: parseFloat(cache.ele.attr('rx')),
-            ry: parseFloat(cache.ele.attr('ry'))
-          }
-          break;
-        case 'circle':
-          this.circle = {
-            cx: parseFloat(cache.ele.attr('cx')),
-            cy: parseFloat(cache.ele.attr('cy')),
-            r: parseFloat(cache.ele.attr('rx'))
-          }
-          break;
-      }
+    // new: function(ele, attr, id) {
+    //   if (svg.created == false) {
+    //     if (!id) {
+    //       id = this.numID;
+    //       this.numID++;
+    //     }
+    //     // the below code simply append to the HTML Document
+    //     $('#editor').html($('#editor').html() + '<' + ele + ' id=' + id + '/>');
+    //     this.id = '#' + id;
+    //     cache.ele = id;
+    //     this.created = true;
+    //     this.finished = false;
+    //   }
+    //   attr['stroke-width'] = tool.strokeWidth;
+    //   attr.stroke = tool.stroke;
+    //   if (ele != 'line' && ele != 'polyline') {
+    //     attr['paint-order'] = tool.paintOrder;
+    //   }
+    //   $(this.id).attr(attr);
+    // },
 
-      var scale = [1,1];
-      var rotate = 0;
-      if (cache.ele.attr('transform')) {
-        var propVal = cache.ele.attr('transform').replace(')','').split(' ');
-        for (var i = 0; i < propVal.length; i++) {
-          if (propVal[i].startsWith('scale')) {
-            var scaleX = parseFloat(propVal[i].split('(')[1].split(',')[0]);
-            var scaleY = parseFloat(propVal[i].split('(')[1].split(',')[1]);
-            scale = [scaleX,scaleY];
-          }
-          if (propVal[i].startsWith('rotate')) {
-            rotate = parseFloat(propVal[i].split('(')[1].split('deg')[0]);
-          }
-        }
-      }
-      this.initial = { // Stores x, y, etc. values so that it can be accessed later to calculate transformations
-        x: cache.ele[0].getBoundingClientRect().left,
-        y: cache.ele[0].getBoundingClientRect().top,
-        width: cache.ele[0].getBoundingClientRect().width,
-        height: cache.ele[0].getBoundingClientRect().height,
-        scale: scale,
-        rotate: rotate,
-        preScaleH: Math.abs(cache.origSelectArea.height - cache.origSelectArea.height/scale[1]),
-        preScaleW: Math.abs(cache.origSelectArea.width - cache.origSelectArea.width/scale[0])
-      }
-      this.newScale = scale;
-    },
     resize: function(axis) {
       var selection = $('.selection')[0].getBoundingClientRect();
       var rightHandle = pressed.handle.includes('right'); // Check if a handle on the right is being pressed
@@ -422,6 +354,76 @@ var svg = {
       }
       draw.selection(cache.ele);
       this.previewMove();
+    },
+    
+    storeAttr: function() {
+      cache.origSelectArea = {
+        x: $('.selection')[0].getBoundingClientRect().left,
+        y: $('.selection')[0].getBoundingClientRect().top,
+        width: $('.selection')[0].getBoundingClientRect().width,
+        height: $('.selection')[0].getBoundingClientRect().height
+      }
+      this.type = cache.ele[0].outerHTML.split(' ')[0].replace('<','');
+      switch (this.type) { // Stores element-specific data for later manipulation
+        case 'line':
+          this.line = {
+            x1: parseFloat(cache.ele.attr('x1')),
+            x2: parseFloat(cache.ele.attr('x2')),
+            y1: parseFloat(cache.ele.attr('y1')),
+            y2: parseFloat(cache.ele.attr('y2'))
+          }
+          break;
+        case 'rect':
+          this.rect = {
+            x: parseFloat(cache.ele.attr('x')),
+            y: parseFloat(cache.ele.attr('y')),
+            width: parseFloat(cache.ele.attr('width')),
+            height: parseFloat(cache.ele.attr('height'))
+          }
+          break;
+        case 'ellipse':
+          this.ellipse = {
+            cx: parseFloat(cache.ele.attr('cx')),
+            cy: parseFloat(cache.ele.attr('cy')),
+            rx: parseFloat(cache.ele.attr('rx')),
+            ry: parseFloat(cache.ele.attr('ry'))
+          }
+          break;
+        case 'circle':
+          this.circle = {
+            cx: parseFloat(cache.ele.attr('cx')),
+            cy: parseFloat(cache.ele.attr('cy')),
+            r: parseFloat(cache.ele.attr('rx'))
+          }
+          break;
+      }
+
+      var scale = [1,1];
+      var rotate = 0;
+      if (cache.ele.attr('transform')) {
+        var propVal = cache.ele.attr('transform').replace(')','').split(' ');
+        for (var i = 0; i < propVal.length; i++) {
+          if (propVal[i].startsWith('scale')) {
+            var scaleX = parseFloat(propVal[i].split('(')[1].split(',')[0]);
+            var scaleY = parseFloat(propVal[i].split('(')[1].split(',')[1]);
+            scale = [scaleX,scaleY];
+          }
+          if (propVal[i].startsWith('rotate')) {
+            rotate = parseFloat(propVal[i].split('(')[1].split('deg')[0]);
+          }
+        }
+      }
+      this.initial = { // Stores x, y, etc. values so that it can be accessed later to calculate transformations
+        x: cache.ele[0].getBoundingClientRect().left,
+        y: cache.ele[0].getBoundingClientRect().top,
+        width: cache.ele[0].getBoundingClientRect().width,
+        height: cache.ele[0].getBoundingClientRect().height,
+        scale: scale,
+        rotate: rotate,
+        preScaleH: Math.abs(cache.origSelectArea.height - cache.origSelectArea.height/scale[1]),
+        preScaleW: Math.abs(cache.origSelectArea.width - cache.origSelectArea.width/scale[0])
+      }
+      this.newScale = scale;
     },
     previewMove: function() {
       var x1 = (this.initial.x - $('#editor').offset().left + this.initial.width/2)/doc.zoom;
